backend/rescheduler/requirements.txt
django
djangorestframework
ortools
django-cors-headers